requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile ./requirements.in
#
aiohttp==3.8.1
    # via python-binance
aiosignal==1.2.0
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==21.4.0
    # via
    #   aiohttp
    #   automat
    #   twisted
automat==20.2.0
    # via twisted
backports.zoneinfo==0.2.1
    # via
    #   pytz-deprecation-shim
    #   tzlocal
bta-lib==1.0.0
    # via -r ./requirements.in
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.10
    # via
    #   aiohttp
    #   requests
constantly==15.1.0
    # via twisted
dateparser==1.1.0
    # via python-binance
frozenlist==1.3.0
    # via
    #   aiohttp
    #   aiosignal
hyperlink==21.0.0
    # via twisted
idna==3.3
    # via
    #   hyperlink
    #   requests
    #   yarl
incremental==21.3.0
    # via twisted
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
numpy==1.22.1
    # via pandas
pandas==1.4.0
    # via
    #   -r ./requirements.in
    #   bta-lib
python-binance==1.0.15
    # via -r ./requirements.in
python-dateutil==2.8.2
    # via
    #   dateparser
    #   pandas
pytz==2021.3
    # via
    #   dateparser
    #   pandas
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
regex==2022.1.18
    # via dateparser
requests==2.27.1
    # via python-binance
six==1.16.0
    # via
    #   automat
    #   python-binance
    #   python-dateutil
twisted==21.7.0
    # via -r ./requirements.in
typing-extensions==4.0.1
    # via twisted
tzdata==2021.5
    # via pytz-deprecation-shim
tzlocal==4.1
    # via dateparser
ujson==5.1.0
    # via python-binance
urllib3==1.26.8
    # via requests
websockets==9.1
    # via python-binance
yarl==1.7.2
    # via aiohttp
zope.interface==5.4.0
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
